+++ demo-ui/src/components/SideBar.tsx
@@ -10,6 +10,7 @@ function SideBar({updateWallet}: WalletProps) {
         <Nav.Link eventKey="view" className={classes} onClick={updateWallet}>View Wallet</Nav.Link>
         <Nav.Link eventKey="getPubKey" className={classes}>Get Public Key</Nav.Link>
         <Nav.Link eventKey="transfer" className={classes}>Transfer BTC</Nav.Link>
+        <Nav.Link eventKey="reclaim" className={classes}>Reclaim BTC</Nav.Link>
     </Nav>
   }