+++ demo-ui/src/components/TabController.tsx
@@ -1,4 +1,5 @@
 import Transfer from '../views/Transfer';
+import Reclaim from '../views/Reclaim';
 import Wallet from '../views/Wallet';
 import GetPubKey from '../views/GetPubKey';
 import Tab from 'react-bootstrap/Tab';
@@ -21,6 +22,7 @@ function TabController() {
       <Tab.Pane eventKey="view"><Wallet balances={balances} transactions={transactions}/></Tab.Pane>
       <Tab.Pane eventKey="getPubKey"><GetPubKey walletName={walletName}/></Tab.Pane>
       <Tab.Pane eventKey="transfer"><Transfer walletName={walletName}/></Tab.Pane>
+      <Tab.Pane eventKey="reclaim"><Reclaim walletName={walletName}/></Tab.Pane>
     </Tab.Content>
   </Tab.Container> 
   }